refactor(node): Move two model mapping functions to `NpmSupport`
The functions `parseProject()` and `parsePackage()` are used by mutliple
package managers. Move them to `NpmSupport` which contains common code
to account for that.
